+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+[unreleased]
+
+- Exclude deleted posts when pulling posts
+
 ## 1.0.0
 
 - Initial release